clinician-service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
PyJWT==2.8.0
requests==2.31.0
gunicorn==21.2.0
python-decouple==3.8
drf-yasg==1.21.7